old-tests/submission/W3C/domhtml/HTMLObjectElement20.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>object</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/the-object-element.html#dom-object-contentdocument" rel="help">
<link rel="author" title="W3C">
</head>
<body>

<p>
<object data="/support/html/sample.html" height="80" type="text/html" width="550"></object>
</p>
<p>
<object data="/support/images/yellow-rectangle.png" height="50" type="image/png" width="100"></object>
</p>

<script>
test(function () {
 var testNode = window.document.getElementsByTagName("object").item(1);
 assert_equals(testNode.contentDocument, null);
}, "object.contentDocument is null");

test(function () {
 var testNode = window.document.getElementsByTagName("object").item(0);
 assert_not_equals(testNode.contentDocument, null);
 assert_equals(testNode.contentDocument.nodeName, "#document");
}, "object.contentDocument is not null");

</script>
<p>The contentDocument IDL attribute must return the Document object of the active document of the object element's nested browsing context, if it has one; otherwise, it must return null.</p>
<div id="log"></div>
</body></html>
